--- Compiler/Gateways/Gateways.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <LangVersion>10</LangVersion>
-        <TargetFramework>netstandard2.1</TargetFramework>
-        <Nullable>enable</Nullable>
-        <Copyright>R-Koubou</Copyright>
-        <Company>R-Koubou</Company>
-        <Product>KSPCompiler.Gateways</Product>
-        <AssemblyName>KSPCompiler.Gateways</AssemblyName>
-        <RootNamespace>KSPCompiler.Gateways</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Domain\Domain.csproj" />
-    </ItemGroup>
-
-</Project>
