+++ Compiler/Infrastructures/Parser.Antlr.Tests/AstTranslatorTest.cs
@@ -3,6 +3,7 @@ using System.Text;
 using KSPCompiler.Domain.Ast.Node.Blocks;
 using KSPCompiler.Domain.CompilerMessages;
 using KSPCompiler.Externals.Parser.Antlr;
+using KSPCompiler.Infrastructures.Parser.Antlr;
 
 using NUnit.Framework;
