+++ Compiler/Infrastructures/Parser.Antlr/Translators/CSTConverterVisitor_Expression_BInary.cs
@@ -4,8 +4,9 @@ using Antlr4.Runtime;
 
 using KSPCompiler.Domain.Ast.Node;
 using KSPCompiler.Domain.Ast.Node.Expressions;
+using KSPCompiler.Externals.Parser.Antlr;
 
-namespace KSPCompiler.Externals.Parser.Antlr.Translators
+namespace KSPCompiler.Infrastructures.Parser.Antlr.Translators
 {
     // implementation of binary operator node generation
     public partial class CSTConverterVisitor
