+++ Compiler/Infrastructures/Parser.Antlr/Translators/CSTConverterVisitor_Statement.cs
@@ -5,9 +5,10 @@ using Antlr4.Runtime;
 using KSPCompiler.Domain.Ast.Node;
 using KSPCompiler.Domain.Ast.Node.Blocks;
 using KSPCompiler.Domain.Ast.Node.Statements;
-using KSPCompiler.Externals.Parser.Antlr.Translators.Extensions;
+using KSPCompiler.Externals.Parser.Antlr;
+using KSPCompiler.Infrastructures.Parser.Antlr.Translators.Extensions;
 
-namespace KSPCompiler.Externals.Parser.Antlr.Translators
+namespace KSPCompiler.Infrastructures.Parser.Antlr.Translators
 {
     // implementation of the statement node
     public partial class CSTConverterVisitor
